Change email to be sent by virtual thread

+++ src/main/java/com/pedrovisk/proxmox/models/proxmox/Memory.java
@@ -3,13 +3,15 @@ package com.pedrovisk.proxmox.models.proxmox;
 import lombok.*;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 @Builder
 @ToString
 public class Memory{
-    public double total;
-    public double used;
-    public double free;
+    public BigDecimal total;
+    public BigDecimal used;
+    public BigDecimal free;
 }
